Implement CollectorStat, add tests, add the newest mock classes
