--- src/main/java/com/softserve/logstat/collector/ParamType.java
@@ -1,9 +1,0 @@
-package com.softserve.logstat.collector;
-
-public enum ParamType {
-        IP,
-        URL,
-        HTTPMETHOD,
-        HTTPVERSION,
-        RC
-}
