--- src/main/java/com/softserve/logstat/collector/Report.java
@@ -1,4 +1,0 @@
-package com.softserve.logstat.collector;
-
-public interface Report {
-}
