+++ src/main/java/com/softserve/logstat/model/ParamType.java
@@ -1,0 +1,13 @@
+package com.softserve.logstat.model;
+
+/**
+ * Класс-заглушка для тестов.
+ * @author Dmitri Chistiuk
+ */
+public enum ParamType {
+        IP,
+        URL,
+        HTTPMETHOD,
+        HTTPVERSION,
+        RC
+}
