+++ src/test/java/com/softserve/logstat/collector/CollectorStatTest.java
@@ -1,7 +1,11 @@
 package com.softserve.logstat.collector;
 
+import com.softserve.logstat.ArgParser;
+import com.softserve.logstat.model.Command;
 import com.softserve.logstat.model.HTTPMethod;
 import com.softserve.logstat.model.Log;
+import com.softserve.logstat.model.report.ReportStat;
+import com.softserve.logstat.service.collector.CollectorStat;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -11,7 +15,7 @@ import java.time.Month;
 import java.util.ArrayList;
 import java.util.HashMap;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class CollectorStatTest {
     private ArrayList<Log> logs = new ArrayList<>();
@@ -105,7 +109,7 @@ class CollectorStatTest {
         statisticManual.put(log3.getRequest(), 1);
         statisticManual.put(log4.getRequest(), 1);
         statisticManual.put(log5.getRequest(), 1);
-        assertEquals(statisticManual, reporterStat.getStatRes());
+        assertEquals(statisticManual, reporterStat.getRes());
     }
 
     @Test
@@ -123,7 +127,7 @@ class CollectorStatTest {
         statisticManual.put(log3.getIp(), 1);
         statisticManual.put(log4.getIp(), 1);
         statisticManual.put(log5.getIp(), 1);
-        assertEquals(statisticManual, reporterStat.getStatRes());
+        assertEquals(statisticManual, reporterStat.getRes());
     }
 
     @Test
@@ -138,7 +142,7 @@ class CollectorStatTest {
         statisticManual = new HashMap<>();
         statisticManual.put(String.valueOf(log1.getHttpVersion()), 4);
         statisticManual.put(String.valueOf(log5.getHttpVersion()), 2);
-        assertEquals(statisticManual, reporterStat.getStatRes());
+        assertEquals(statisticManual, reporterStat.getRes());
     }
 
     @Test
@@ -154,7 +158,7 @@ class CollectorStatTest {
         statisticManual.put(String.valueOf(log1.getResponseCode()), 4);
         statisticManual.put(String.valueOf(log4.getResponseCode()), 1);
         statisticManual.put(String.valueOf(log5.getResponseCode()), 1);
-        assertEquals(statisticManual, reporterStat.getStatRes());
+        assertEquals(statisticManual, reporterStat.getRes());
     }
     @Test
     @DisplayName("Find -httpMethod statistic(log1,log2,copyLog1 -> GET, log3,log4-> POST log5 -> PUT).")
@@ -169,7 +173,7 @@ class CollectorStatTest {
         statisticManual.put(String.valueOf(log1.getMethod()), 3);
         statisticManual.put(String.valueOf(log3.getMethod()), 2);
         statisticManual.put(String.valueOf(log5.getMethod()), 1);
-        assertEquals(statisticManual, reporterStat.getStatRes());
+        assertEquals(statisticManual, reporterStat.getRes());
     }
     @Test
     @DisplayName("Find -url statistic limiting by size with less(one log is copy).The log5,log3 will not be in result")
@@ -188,7 +192,7 @@ class CollectorStatTest {
         statisticManual.put(copyLog1.getRequest(), 2);
         statisticManual.put(log2.getRequest(), 1);
         statisticManual.put(log4.getRequest(), 1);
-        assertEquals(statisticManual, reporterStat.getStatRes());
+        assertEquals(statisticManual, reporterStat.getRes());
     }
     @Test
     @DisplayName("Find -url statistic limiting by size with more(one log is copy).The log2,log1,log1Copy will not be in result")
@@ -207,7 +211,7 @@ class CollectorStatTest {
         statisticManual.put(log2.getRequest(), 2);
         statisticManual.put(log3.getRequest(), 1);
         statisticManual.put(log5.getRequest(), 1);
-        assertEquals(statisticManual, reporterStat.getStatRes());
+        assertEquals(statisticManual, reporterStat.getRes());
     }
     @Test
     @DisplayName("Find -url statistic limiting by 2 different size(one log is copy).The log5,log2,log1,log1Copy will not be in result")
@@ -228,7 +232,7 @@ class CollectorStatTest {
         statisticManual = new HashMap<>();
         statisticManual.put(log2.getRequest(), 2);
         statisticManual.put(log3.getRequest(), 1);
-        assertEquals(statisticManual, reporterStat.getStatRes());
+        assertEquals(statisticManual, reporterStat.getRes());
     }
     @Test
     @DisplayName("Find -url statistic limiting by size with equals(one log is copy).The log5 will be in result")
@@ -245,7 +249,7 @@ class CollectorStatTest {
         reporterStat = (ReportStat) collectorStat.collect(logs.stream(),command);
         statisticManual = new HashMap<>();
         statisticManual.put(log5.getRequest(), 1);
-        assertEquals(statisticManual, reporterStat.getStatRes());
+        assertEquals(statisticManual, reporterStat.getRes());
     }
     @Test
     @DisplayName("Find -url statistic limiting by size with not equals(one log is copy).The log5 will not be in result")
@@ -265,7 +269,7 @@ class CollectorStatTest {
         statisticManual.put(log2.getRequest(), 1);
         statisticManual.put(log3.getRequest(), 1);
         statisticManual.put(log4.getRequest(), 1);
-        assertEquals(statisticManual, reporterStat.getStatRes());
+        assertEquals(statisticManual, reporterStat.getRes());
     }
     @Test
     @DisplayName("Find -url statistic limiting by time with not equals(one log is copy).The log5 will not be in result")
@@ -285,7 +289,7 @@ class CollectorStatTest {
         statisticManual.put(log2.getRequest(), 1);
         statisticManual.put(log3.getRequest(), 1);
         statisticManual.put(log4.getRequest(), 1);
-        assertEquals(statisticManual, reporterStat.getStatRes());
+        assertEquals(statisticManual, reporterStat.getRes());
     }
     @Test
     @DisplayName("Find -url statistic limiting by time with equals(one log is copy).The log5 will not be in result")
@@ -302,7 +306,7 @@ class CollectorStatTest {
         reporterStat = (ReportStat) collectorStat.collect(logs.stream(),command);
         statisticManual = new HashMap<>();
         statisticManual.put(log5.getRequest(), 1);
-        assertEquals(statisticManual, reporterStat.getStatRes());
+        assertEquals(statisticManual, reporterStat.getRes());
     }
     @Test
     @DisplayName("Find -url statistic limiting by time with less(one log is copy).The log5,log4 will not be in result")
@@ -321,7 +325,7 @@ class CollectorStatTest {
         statisticManual.put(log1.getRequest(), 2);
         statisticManual.put(log2.getRequest(), 1);
         statisticManual.put(log3.getRequest(), 1);
-        assertEquals(statisticManual, reporterStat.getStatRes());
+        assertEquals(statisticManual, reporterStat.getRes());
     }
     @Test
     @DisplayName("Find -url statistic limiting by time with more(one log is copy).The log5,log4 will be in result")
@@ -339,7 +343,7 @@ class CollectorStatTest {
         statisticManual = new HashMap<>();
         statisticManual.put(log5.getRequest(), 1);
         statisticManual.put(log4.getRequest(), 1);
-        assertEquals(statisticManual, reporterStat.getStatRes());
+        assertEquals(statisticManual, reporterStat.getRes());
     }
     @Test
     @DisplayName("Find -url statistic limiting by time with between(one log is copy).The log5,log4 will be in result")
@@ -358,6 +362,6 @@ class CollectorStatTest {
         statisticManual = new HashMap<>();
         statisticManual.put(log5.getRequest(), 1);
         statisticManual.put(log4.getRequest(), 1);
-        assertEquals(statisticManual, reporterStat.getStatRes());
+        assertEquals(statisticManual, reporterStat.getRes());
     }
 }
